chore(phpstan): enforce level 7 on the compiler module (#2400)
Hold the correctness-critical compiler pipeline (lexer/parser/reader/
analyzer/emitter) to PHPStan level 7 via a scoped phpstan-compiler.neon
that inherits the shared level-6 baseline. Wired into composer
test-quality and the CI type-checker job.

Fixes the 19 level-7 findings this surfaced:
- MultiFnNode: guard empty arity arrays before min/max
- StringParser: cast hexdec() to int; correct matches param type
- CallInliner/PureLiteralDetector: align list vs array node types
- DefInterfaceSymbol: build a typed list<Symbol> of method arguments
- NsSymbol: array<int,mixed> require-entry element types
- ByRefLocalCollector/BodyConstantScanner/LocalVarReferences: honest
  array<int,AbstractNode> children() return types
- SourceMapConsumer: null-safe min over line mappings
- Phel::list(): accept array<int,mixed> (matches the factory)

+++ src/php/Compiler/Domain/Parser/ExpressionParser/StringParser.php
@@ -52,7 +52,7 @@ final class StringParser
             }
 
             if ($str[0] === 'x' || $str[0] === 'X') {
-                return chr(hexdec(substr((string) $str, 1)));
+                return chr((int) hexdec(substr((string) $str, 1)));
             }
 
             if ($str[0] === 'u') {
@@ -76,7 +76,7 @@ final class StringParser
     }
 
     /**
-     * @param list<string> $matches
+     * @param array<int, string> $matches
      *
      * @throws StringParserException
      */
@@ -88,7 +88,7 @@ final class StringParser
             $hexCodepoint = $matches[self::FIXED_UNICODE_ESCAPE_MATCH] ?? '';
         }
 
-        return $this->codePointToUtf8(hexdec($hexCodepoint));
+        return $this->codePointToUtf8((int) hexdec($hexCodepoint));
     }
 
     /**
